+++ hooks/configurations/unzipAndCopyConfigurations.js
@@ -5,6 +5,8 @@ var AdmZip = require("adm-zip");
 
 var utils = require("./utilities");
 
+var utilsApp = require("../utilities");
+
 var constants = {
   googleServices: "google-services"
 };
@@ -53,13 +55,31 @@ module.exports = function(context) {
   var sourceFilePath = path.join(targetPath, fileName);
   var destFilePath = path.join(context.opts.plugin.dir, fileName);
 
-  utils.copyFromSourceToDestPath(defer, sourceFilePath, destFilePath);
+  var androidPath =  "platforms/android/app";
+  var iOSPath = "platforms/ios/" + utilsApp.getAppName(context) + "/Resources";
+
+  var completeFilePath;
+
+  var isAndroid = platform.localeCompare("android");
+  var isIOS = platform.localeCompare("ios");
+
+  if(isAndroid == 0){ //android code
+    completeFilePath = path.join(context.opts.projectRoot, androidPath);
+  }
+  else if(isIOS == 0){ //iOS code
+    completeFilePath = path.join(context.opts.projectRoot, iOSPath);
+  }
+
+  if(!utils.checkIfFolderExists(destFilePath)){
+    utils.copyFromSourceToDestPath(defer, sourceFilePath, destFilePath);
+  }
 
   if (cordovaAbove7) {
-    var destPath = path.join(context.opts.projectRoot, "platforms", platform, "app");
-    if (utils.checkIfFolderExists(destPath)) {
-      var destFilePath = path.join(destPath, fileName);
-      utils.copyFromSourceToDestPath(defer, sourceFilePath, destFilePath);
+    if (utils.checkIfFolderExists(completeFilePath)) {
+      var destFilePath = path.join(completeFilePath, fileName);
+      if(!utils.checkIfFolderExists(destFilePath)){
+        utils.copyFromSourceToDestPath(defer, sourceFilePath, destFilePath);
+      }
     }
   }